+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 # What's new:
 - Fixed a rendering bug with bells
+
+# Patch 1:
+- Added compatiblity with Every Compat
+- Added compatibility with Carved Wood